+++ packages/jsts/src/rules/helpers/index.ts
@@ -32,6 +32,7 @@ export * from './location.js';
 export * from './lva.js';
 export * from './mocha.js';
 export * from './module.js';
+export * from './module-ts.js';
 export * from './package-json.js';
 export * from './quickfix.js';
 export * from './reaching-definitions.js';
